Set STREAMLIT location as in the instructions
The instructions tell the user to upload all files to the root of the
stage. The setup_script creates STREAMLIT from /src/ subdirectory.

Without this change, application is created but fails to start.

Alternative is to instruct the user to upload the files to a
subdirectory but that would involve specifying `src` as the path when
uploading files.